Type for start_position and end_position changed to size_t. Refs #142
git-svn-id: http://alma-dms/svn/sakura/trunk@254 9a79bd28-2a0b-46f8-beb7-14fc758f186d

+++ libsakura/src/interpolation.cc
@@ -458,8 +458,8 @@ LIBSAKURA_SYMBOL(InterpolationMethod) interpolation_method,
 						sizeof(size_t) * num_base));
 
 		// Locate each element in x_base against x_interpolated
-		int start_position = 0;
-		int end_position = static_cast<int>(num_interpolated) - 1;
+		size_t start_position = 0;
+		size_t end_position = num_interpolated - 1;
 		for (size_t i = 0; i < num_base; ++i) {
 			location_base[i] = this->Locate(start_position, end_position,
 					num_interpolated, x_interpolated_work, x_base_work[i]);
